refactor: 회원가입 API 수정
