+++ src/main/java/com/codeit/sprint/team3/backend/auth/adapter/in/web/RegisterUserController.java
@@ -1,6 +1,6 @@
 package com.codeit.sprint.team3.backend.auth.adapter.in.web;
 
-import com.codeit.sprint.team3.backend.auth.application.port.in.RegisterUserCommand;
+import com.codeit.sprint.team3.backend.auth.application.port.in.RegisterUserRequest;
 import com.codeit.sprint.team3.backend.auth.application.port.in.RegisterUserUseCase;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -19,18 +19,10 @@ public class RegisterUserController {
 
     @PostMapping("/signup")
     public ResponseEntity<Map<String, String>> register(
-            @Valid //검증 실패 시 MethodArgumentNotValidException 발생 -> ExceptionHandler 에서 처리
-            @RequestBody
-            RegisterUserRequest request) {
-        RegisterUserCommand command = new RegisterUserCommand(
-                request.getName(),
-                request.getEmail(),
-                request.getPassword(),
-                request.getNickName(),
-                request.getDescription()
-        );
+            @Valid @RequestBody RegisterUserRequest request
+    ) {
 
-        registerUserUseCase.register(command);
+        registerUserUseCase.register(request);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(Map.of("message", "사용자 생성 성공"));
     }
