+++ src/main/java/com/codeit/sprint/team3/backend/auth/application/port/in/RegisterUserRequest.java
@@ -1,24 +1,26 @@
-package com.codeit.sprint.team3.backend.auth.adapter.in.web;
+package com.codeit.sprint.team3.backend.auth.application.port.in;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
-@Getter @Setter
+@Data
+@Builder
+@AllArgsConstructor
 public class RegisterUserRequest {
     @NotBlank(message = "이름은 필수값입니다.")
-    private String name;
+    private final String name;
 
     @Email(message = "유효한 이메일 주소를 입력하세요")
     @NotBlank(message = "이메일은 필수값입니다.")
-    private String email;
+    private final String email;
 
     @NotBlank(message = "비밀번호는 필수값입니다.")
-    private String password;
+    private final String password;
 
     @NotBlank(message = "닉네임은 필수값입니다.")
-    private String nickName;
+    private final String nickname;
 
-    private String description;
+    private final String description;
+
 }
