+++ src/main/java/com/codeit/sprint/team3/backend/auth/application/port/in/RegisterUserUseCase.java
@@ -1,5 +1,5 @@
 package com.codeit.sprint.team3.backend.auth.application.port.in;
 
 public interface RegisterUserUseCase {
-    Long register(RegisterUserCommand command);
+    Long register(RegisterUserRequest command);
 }
